no reason to use float here
git-svn-id: https://svn.salilab.org/imp/trunk@16762 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/statistics/src/Histogram.cpp
@@ -55,7 +55,7 @@ void Histogram::show(std::ostream& out) const {
 double Histogram::get_top(double percentage) const {
   IMP_INTERNAL_CHECK(!((percentage<0.)||(percentage>1.)),
                      "The input number is not a percentage\n");
-  float stop_count = get_total_count()*percentage;
+  double stop_count = get_total_count()*percentage;
   int partial_count=0;
   for(unsigned int i=0;i<freq_.size();i++) {
     partial_count += freq_[i];
